move textsearch tests to jobutil
The only thing in the textsearch package is a test, so I moved that test
into the jobutil package becuase the test covers both zoekt and searcher
jobs.
