--- internal/search/textsearch/textsearch.go
@@ -1,1 +1,0 @@
-package textsearch
